Make RowCoder a trival sub-class of SchemaCoder<Row>
... so that it can be used as the java implementation of
beam:coder:row:v1 in a later change

All of its remaining functionality has been moved to SchemaCoder and
RowCoderGenerator.
